smooooooth hover stuff for 80+ courses
