+++ frontend/src/components/WeekSchedule.js
@@ -1,22 +1,13 @@
 import React, { useCallback, useMemo } from 'react';
-import { OverlayTrigger, Popover } from 'react-bootstrap';
 import moment from 'moment';
 import './WeekSchedule.css';
 import { momentLocalizer } from 'react-big-calendar';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
-import LinesEllipsis from 'react-lines-ellipsis';
-import responsiveHOC from 'react-lines-ellipsis/lib/responsiveHOC';
 import { Calendar } from 'react-big-calendar';
-import { StyledPopover } from './StyledComponents';
 import styled from 'styled-components';
+import CalendarEvent from './CalendarEvent';
 
-const ResponsiveEllipsis = responsiveHOC()(LinesEllipsis);
 const localizer = momentLocalizer(moment);
-
-const event_style = {
-  width: '100%',
-  height: '100%',
-};
 
 // Calendar for worksheet
 const StyledCalendar = styled(Calendar)`
@@ -108,97 +99,20 @@ function WeekSchedule({ showModal, courses, hover_course, hidden_courses }) {
     return [earliest, latest, parsedCourses];
   }, []);
 
-  // Render popover that contains title, description, and requirements when hovering over course
-  const renderTitlePopover = useCallback((props, course) => {
-    return (
-      <StyledPopover {...props} id="title_popover">
-        <Popover.Title>
-          <strong>{course.title}</strong>
-        </Popover.Title>
-        <Popover.Content>
-          {course.description
-            ? course.description.length <= 300
-              ? course.description
-              : course.description.slice(0, 300) + '...'
-            : 'no description'}
-          <br />
-          <div className="text-danger">
-            {course.requirements &&
-              (course.requirements.length <= 250
-                ? course.requirements
-                : course.requirements.slice(0, 250) + '...')}
-          </div>
-        </Popover.Content>
-      </StyledPopover>
-    );
-  }, []);
-
-  // Render the custom data displayed on each calendar event
-  const customEvent = useCallback(
-    (event) => {
-      return (
-        <OverlayTrigger
-          // Course info that appears on hover
-          placement="right"
-          overlay={(props) => renderTitlePopover(props, event.event.listing)}
-          // Have a 500ms delay before showing popover so it only pops up when user wants it to
-          delay={{ show: 300, hide: 100 }}
-        >
-          <div
-            style={event_style}
-            // onMouseEnter={() => setHoverCourse(event.event.listing)}
-            // onMouseLeave={() => setHoverCourse(null)}
-          >
-            <strong>{event.title}</strong>
-            <br />
-            <span style={{ fontSize: '12px' }}>
-              <ResponsiveEllipsis
-                style={{ whiteSpace: 'pre-wrap' }}
-                text={event.event.listing.title}
-                maxLine={'2'}
-                basedOn="words"
-              />
-            </span>
-            <small className="location_text">
-              {event.event.listing.locations_summary}
-            </small>
-          </div>
-        </OverlayTrigger>
-      );
-    },
-    // [setHoverCourse]
-    [renderTitlePopover]
-  );
-
   // Custom styling for the calendar events
   const eventStyleGetter = useCallback(
     (event) => {
-      const border = '1)';
-      let style;
+      let style = {
+        backgroundColor: event.listing.color,
+        borderColor: event.listing.border,
+        borderWidth: '2px',
+      };
       if (hidden_courses[event.listing.crn]) {
         style = { display: 'none' };
       } else if (hover_course && hover_course === event.listing.crn) {
-        style = {
-          backgroundColor: event.listing.color.concat('1)'),
-          borderColor: event.listing.color.concat(border),
-          borderWidth: '2px',
-          filter: 'saturate(130%)',
-          transform: 'scale(1.03)',
-          zIndex: 69,
-        };
+        style.zIndex = 2;
       } else if (hover_course) {
-        style = {
-          backgroundColor: event.listing.color.concat('.85)'),
-          borderColor: event.listing.color.concat(border),
-          borderWidth: '2px',
-          opacity: '40%',
-        };
-      } else {
-        style = {
-          backgroundColor: event.listing.color.concat('.85)'),
-          borderColor: event.listing.color.concat(border),
-          borderWidth: '2px',
-        };
+        style.opacity = '40%';
       }
       return {
         style: style,
@@ -222,10 +136,6 @@ function WeekSchedule({ showModal, courses, hover_course, hidden_courses }) {
       ? ret_values[1].toDate()
       : moment().hour(18).minute(0).toDate();
   }, [ret_values]);
-
-  const customEventMemo = useMemo(() => {
-    return { event: customEvent };
-  }, [customEvent]);
 
   const onSelectEventCallback = useCallback(
     (event) => {
@@ -254,7 +164,7 @@ function WeekSchedule({ showModal, courses, hover_course, hidden_courses }) {
       localizer={localizer}
       toolbar={false}
       onSelectEvent={onSelectEventCallback}
-      components={customEventMemo}
+      components={{ event: CalendarEvent }}
       eventPropGetter={eventPropGetterCallback}
       // Display Mon, Tue, Wed, ... at the top
       formats={{
